Fixed notes and renamed the packages
